Improve definition of PDS::DataAdapter::Base
Add additional yardoc definition for return types and errors raised.

Set appropriate arg defaults for DataAdapter::Base.

Rename `filter` to `filters`, to reflect that this accepts an array of
values (plural).

When `filters` is available, it is optional. Set the default for the base
class to an empty array, as the suggestion for how to implementing
classes should handle it.

When `resources` is available, it is required. Do not provide a default
value for this parameter in the base class.