Move PG models to data adapter class
We want to be able to use other ORM model types for different databases,
so this commit ensures that will be easy by moving the model definitions
for PostgreSQL to a PostgreSQL-specific location. Because these models
are very small, it's easy to simply define them inside the PostgreSQL
Data Adapter class itself for now.

This also allows us to avoid loading ActiveRecord unless/until the user
configuration for PostgreSQL is being processed. This means the app does
not load ActiveRecord if it is not needed, such as when the mock data
adapter is being used, or in a possible future state when a different
ORM is used instead (such as for Cassandra or Mongo).