Standardize on flags for IDs in upsert
Use flags for IDs, not positional arguments. Hiera Data doesn't
intuitively work to use positional arguments due to the composite key
(we don't want users to have to remember positional ordering), and we
want all commands to have the same UX. Therefore, we should start moving
to standardize on using flags for all parameters, even required ones,
instead of positional arguments.

This commit switches upsert in node and user to use flags for IDs, not
args[0].