Use name "Hiera Datum" in API spec
The use of "Hiera Datum" replaces use of "Hiera Value" to refer to an
individual resource of type Hiera Data. This change is primarily just in
the API spec file.

Using Hiera Datum instead of Hiera Value for this avoids potential
confusion coming from the fact that a single Hiera Datum resource has a
field called "value".