Work around committee limitation
Committee can't process request bodies that aren't JSON objects. That
is, it can't process array bodies. So, because validation is important
to us, modify the API to use an object wrapper for our POST requests.
This is probably a good idea anyway, it lets us more easily create
different kinds of POSTS if we need to in the future.