Fix two rpm build problems
Problem 1:
The pe-postgresql11 dependency didn't specify a version. This let the
package be installed on systems with an older-than-required version of
PostgreSQL, which caused problems.

Problem 2:
The rpm inadvertently included a hardlink, which was likely to span
filesystems. This was a total accident. Fix build not to include
*.example file which was causing the problem and accidental hardlink.

Fixes #53

Bonus problem:
It should be possible to build an RPM on an OSX workstation, for
testing. The Makefile had a redundant unguarded dependency though which
make Make try to use yum to install golang, which won't work, obvs. Was
a bug. Fixed it.